+++ test/hooks/index.test.js
@@ -7,6 +7,8 @@ const { lookupSchool } = require('../../src/hooks');
 const appPromise = require('../../src/app');
 const { createTestUser, cleanup } = require('../services/helpers/testObjects')(appPromise);
 
+const { hasPermission } = require('../../src/hooks/index');
+
 describe('#lookupSchool', () => {
 	let app;
 
@@ -47,4 +49,31 @@ describe('#lookupSchool', () => {
 	});
 
 	after(cleanup);
+});
+
+describe('check permission', () => {
+	it('should fail if permissions operator is wrong', () => {
+		const inputPermission = 'RANDOM';
+		const operator = 'FOO';
+		try {
+			hasPermission(inputPermission, operator);
+			throw new Error('should have failed');
+		} catch (err) {
+			expect(err.message).to.not.equal('should have failed');
+			expect(err.code).to.equal(403);
+			expect(err.message).to.equal('Invalid function call');
+		}
+	});
+
+	it('should fail if permission is not set', () => {
+		const inputPermission = '';
+		try {
+			hasPermission(inputPermission);
+			throw new Error('should have failed');
+		} catch (err) {
+			expect(err.message).to.not.equal('should have failed');
+			expect(err.code).to.equal(403);
+			expect(err.message).to.equal('Invalid function call');
+		}
+	});
 });